+++ packages/frontend/src/api/csv.ts
@@ -18,6 +18,7 @@ export const downloadCsv = async ({
     columnOrder,
     customLabels,
     hiddenFields,
+    chartName,
 }: {
     projectUuid: string;
     tableId: string;
@@ -28,6 +29,7 @@ export const downloadCsv = async ({
     columnOrder: string[];
     customLabels?: Record<string, string>;
     hiddenFields?: string[];
+    chartName?: string;
 }) => {
     const timezoneFixQuery = {
         ...query,
@@ -44,6 +46,7 @@ export const downloadCsv = async ({
             customLabels,
             columnOrder,
             hiddenFields,
+            chartName,
         }),
     });
 };