+++ packages/frontend/src/components/DashboardTiles/TileForms/MarkdownTileForm.tsx
@@ -1,4 +1,9 @@
-import { type DashboardMarkdownTileProperties } from '@lightdash/common';
+import {
+    HTML_SANITIZE_MARKDOWN_TILE_RULES,
+    sanitizeHtml,
+    type DashboardMarkdownTile,
+    type DashboardMarkdownTileProperties,
+} from '@lightdash/common';
 import { Stack, TextInput } from '@mantine/core';
 import { type UseFormReturnType } from '@mantine/form';
 import MDEditor from '@uiw/react-md-editor';
@@ -6,6 +11,16 @@ import MDEditor from '@uiw/react-md-editor';
 interface MarkdownTileFormProps {
     form: UseFormReturnType<DashboardMarkdownTileProperties['properties']>;
 }
+
+/**
+ * Helper that can be used as a value transformer with Mantine's `useForm` hook.
+ */
+export const markdownTileContentTransform = (
+    values: DashboardMarkdownTile['properties'],
+) => ({
+    ...values,
+    content: sanitizeHtml(values.content, HTML_SANITIZE_MARKDOWN_TILE_RULES),
+});
 
 const MarkdownTileForm = ({ form }: MarkdownTileFormProps) => (
     <Stack spacing="md">